Move ProcessUtils to Utilities namespace
